// src/src/types/boq.ts
export interface Signature {
  type: "proposer" | "customer" | "witness";
  name: string;
  position?: string;
  signatureUrl?: string;
  date?: string;
}